agent-zero-backup/workdir/sentient-api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
pydantic[email]>=2.6
slowapi>=0.1.9
dnspython>=2.6
python-multipart>=0.0.9
